python/requirements.txt
google-api-python-client
pymongo
textblob
python-dotenv

